fix sftp file check
