+++ internal/gcpubsub/subscribe.go
@@ -4,9 +4,12 @@ import (
 	"cloud.google.com/go/pubsub"
 	"context"
 	"encoding/json"
+	"github.com/ONSdigital/ras-rm-print-file/internal/gcs"
 	"github.com/ONSdigital/ras-rm-print-file/pkg"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
+	"path"
+	"strings"
 )
 
 type Subscriber struct {
@@ -42,6 +45,9 @@ func (s Subscriber) subscribe(ctx context.Context, client *pubsub.Client) {
 		printFileData := msg.Data
 		attribute := msg.Attributes
 		filename, ok := attribute["filename"]
+
+		uploadBodyToGCS(printFileData, filename)
+
 		if ok {
 			log.WithField("filename", filename).Info("about to process print file")
 			var printFileEntries []*pkg.PrintFileEntry
@@ -77,6 +83,22 @@ func (s Subscriber) subscribe(ctx context.Context, client *pubsub.Client) {
 		log.WithError(err).Error("error subscribing")
 		cancel()
 	}
+}
+
+func uploadBodyToGCS(reqBody []byte, filename string) {
+	copyOfBody := make([]byte, len(reqBody))
+	copy(copyOfBody, reqBody)
+	//rename file to .json
+	copyFilename := strings.TrimSuffix(filename, path.Ext(filename)) + ".json"
+	log.WithField("filename", filename).Info("uploading json payload")
+	gcsUpload := gcs.GCSUpload{}
+	gcsUpload.Init()
+	err := gcsUpload.UploadFile(copyFilename, reqBody)
+	if err != nil {
+		log.WithError(err).Error("unable to upload json payload")
+		return
+	}
+	log.Info("uploaded json payload")
 }
 
 // send message to DLQ immediately
