+++ internal/web/rest.go
@@ -3,7 +3,6 @@ package web
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/ONSdigital/ras-rm-print-file/internal/gcs"
 	"github.com/ONSdigital/ras-rm-print-file/internal/processor"
 	"github.com/ONSdigital/ras-rm-print-file/pkg"
 	"github.com/gorilla/mux"
@@ -29,9 +28,6 @@ func Print(w http.ResponseWriter, r *http.Request) {
 			log.WithError(err).Error("unable to read body")
 			w.WriteHeader(http.StatusInternalServerError)
 		}
-
-
-
 		log.WithField("reqBody", string(reqBody)).Debug("body of request")
 		vars := mux.Vars(r)
 		filename := vars["filename"]
@@ -41,7 +37,6 @@ func Print(w http.ResponseWriter, r *http.Request) {
 		}
 		log.WithField("filename", filename).Info("received request to print file")
 		var printFileEntries []*pkg.PrintFileEntry
-
 		err = json.Unmarshal(reqBody, &printFileEntries)
 		if err != nil {
 			log.WithError(err).Error("unable to marshall json payload")
@@ -54,8 +49,6 @@ func Print(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusAccepted)
 		resp, _ := json.Marshal(printFile)
 		log.WithField("resp", string(resp)).Debug("about to process")
-		uploadBodyToGCS(reqBody, filename)
-
 		//spawn a process to process the print file
 
 		go processor.CreateAndProcess(filename, &printFile)
@@ -64,15 +57,6 @@ func Print(w http.ResponseWriter, r *http.Request) {
 		log.Info("print - method not allowed")
 		fmt.Fprintf(w, "Only POST methods are supported.")
 	}
-}
-
-func uploadBodyToGCS(reqBody []byte, filename string) {
-	copyOfBody := make([]byte, len(reqBody))
-	copy(copyOfBody, reqBody)
-	copyFilename := filename + ".json"
-	gcsUpload := gcs.GCSUpload{}
-	gcsUpload.Init()
-	gcsUpload.UploadFile(copyFilename, copyOfBody)
 }
 
 func Alive(w http.ResponseWriter, r *http.Request) {
